Housecleaning
Updated the conditional loops to be more pythonic and readable, also changed the open/write calls to work with Python 2.7 (version I'm using,t hese bits aren't that important and can be reverted back).

Want to make it more scalable for more etymologies, i.e. have for x in languageList: print ... kind of statements, so that you don't need to update the code to accommodate new languages, just the dictionary.

Taking baby steps towards this by creating a dictionary with languages and word counts, and with colours.